load font on focus
